refactor content provider A custom ContentProvider now abstracts ContentContext.Provider This means the provider can be more explicit with props to feed the context
